Use sqrt instead of sqrtf

+++ src/math/VectorHelpers.h
@@ -8,7 +8,7 @@ namespace rt {
 
     template<typename T>
     float norm(const T& vec) {
-        float sum = vec[0]*vec[0] + vec[1]*vec[1] + vec[2]*vec[2];
-        return std::sqrtf(sum);
+        double sum = vec[0]*vec[0] + vec[1]*vec[1] + vec[2]*vec[2];
+        return std::sqrt(sum);
     }
 }
